bigdata_3/requirements.txt
pyspark==3.5.6
kafka-python==2.0.2
pandas==2.2.2
openpyxl==3.1.2